compelte

+++ m1-w3d4-polymorphism-exercises/src/main/java/com/techelevator/postage/IDeliveryDriver.java
@@ -3,7 +3,9 @@
 
 package com.techelevator.postage;
 
+import java.math.BigDecimal;
+
 public interface IDeliveryDriver {
 
-	double calculateRate(int distance, int weight);
+	BigDecimal calculateRate(BigDecimal distance, int weight);
 }
